+++ src/main/java/com/example/ehe_server/service/intf/auth/PasswordResetRequestServiceInterface.java
@@ -2,5 +2,5 @@ package com.example.ehe_server.service.intf.auth;
 
 public interface PasswordResetRequestServiceInterface {
     void requestPasswordResetForUnauthenticatedUser(String email);
-    void requestPasswordResetForAuthenticatedUser(Long userId);
+    void requestPasswordResetForAuthenticatedUser(Integer userId);
 }
