+++ src/main/java/com/example/ehe_server/service/intf/user/ApiKeyRemovalServiceInterface.java
@@ -1,5 +1,5 @@
 package com.example.ehe_server.service.intf.user;
 
 public interface ApiKeyRemovalServiceInterface {
-    void removeApiKey(Long userId, Integer apiKeyId);
+    void removeApiKey(Integer userId, Integer apiKeyId);
 }
